requirements.txt
streamlit==1.47.1
pillow==11.3.0
PyMuPDF==1.26.3
pytesseract==0.3.13
opencv-python==4.12.0.88
numpy==2.2.6
